integration: extract test_prepared_statement_col_specs()
The test is extracted into `statements/prepared.rs`.
